[Refactor] type 폴더 따로 빼버리기
