+++ frontend/src/Component/Detail/InfomationCard.tsx
@@ -15,7 +15,7 @@ import AccessTimeFilledIcon from '@mui/icons-material/AccessTimeFilled';
 import ApartmentIcon from '@mui/icons-material/Apartment';
 import EmojiEventsIcon from '@mui/icons-material/EmojiEvents';
 import InboxIcon from '@mui/icons-material/Inbox';
-import { IInfomation } from '../../Util/Type';
+import { IInfomation } from '../../Types/Type';
 import { BASIC_INFO_TYPE, EXTRA_INFO_TYPE } from '../../Util/Constant';
 
 interface IInfomationCard {
