+++ frontend/src/Component/Main/ShareCard.tsx
@@ -1,9 +1,9 @@
-import { Avatar, Box, Card, CardMedia, Chip, Typography } from '@mui/material';
+import { Box, Card, CardMedia, Typography } from '@mui/material';
 import { useCallback } from 'react';
 import { useDispatch } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
 import { selectOnePost } from '../../Store/Action/postAction';
-import { IData } from '../../Util/Type';
+import { IData } from '../../Types/Type';
 
 interface IShareCard {
   data: IData;
@@ -31,14 +31,14 @@ const ShareCard: React.FC<IShareCard> = ({ data }: IShareCard) => {
       <Box className='card-content-container'>
         <Typography>{data.title}</Typography>
         <Box className='card-label-container'>
-          {data.label.map((label, idx) => (
+          {/* {data.label.map((label, idx) => (
             <Chip key={idx} className='card-label' label={label} />
-          ))}
+          ))} */}
         </Box>
       </Box>
       <Box className='card-user-container'>
-        <Avatar alt='user' src={data.User.img} />
-        <Typography>{data.User.name}</Typography>
+        {/* <Avatar alt='user' src={data.User.img} /> */}
+        {/* <Typography>{data.User.name}</Typography> */}
       </Box>
     </Card>
   );
