+++ frontend/src/Component/Main/index.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useLayoutEffect, useState } from 'react';
 import { useDispatch } from 'react-redux';
 import { getAllPost } from '../../Store/Action/postAction';
-import { IData } from '../../Util/Type';
+import { IData } from '../../Types/Type';
 import Filter from './Filter';
 import ShareTab from './ShareTab';
 
@@ -15,9 +15,11 @@ const Main: React.FC = () => {
   const dispatch = useDispatch();
 
   const setInitialPosts = useCallback(async () => {
-    const res = await dispatch(getAllPost({ limit: 1, offset: 1 }));
+    const res = await dispatch(getAllPost({ limit: 3, offset: 0 }));
+
     if (res.payload.status === 200) {
-      setTalentData(res.payload.data);
+      const posts = res.payload.data.data as IData[];
+      setTalentData(posts);
     }
   }, [dispatch]);
