+++ frontend/src/Types/UserType.ts
@@ -1,0 +1,7 @@
+// TODO: User DTO
+export interface User {}
+
+export interface LoginRequeset {
+  email: string;
+  password: string;
+}
